+++ cmd/frontend/internal/graphqlbackend/users.go
@@ -2,26 +2,83 @@ package graphqlbackend
 
 import (
 	"context"
+	"errors"
+	"fmt"
+	"strconv"
+	"sync"
 
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/app/envvar"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/backend"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/db"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/pkg/types"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/pkg/useractivity"
 )
 
 func (r *siteResolver) Users(args *struct {
 	connectionArgs
-	Query *string
+	Query        *string
+	ActivePeriod *string
 }) *userConnectionResolver {
 	var opt db.UsersListOptions
 	if args.Query != nil {
 		opt.Query = *args.Query
 	}
 	args.connectionArgs.set(&opt.LimitOffset)
-	return &userConnectionResolver{opt: opt}
+	return &userConnectionResolver{opt: opt, activePeriod: args.ActivePeriod}
 }
 
 type userConnectionResolver struct {
-	opt db.UsersListOptions
+	opt          db.UsersListOptions
+	activePeriod *string
+
+	// cache results because they are used by multiple fields
+	once       sync.Once
+	users      []*types.User
+	totalCount int
+	err        error
+}
+
+// compute caches results from the more expensive user list creation that occurs when activePeriod
+// is set to a specific length of time.
+//
+// Because user activity data isn't stored in PostgreSQL (but rather in
+// Redis), adding this parameter requires accessing a second data store.
+func (r *userConnectionResolver) compute(ctx context.Context) ([]*types.User, int, error) {
+	if r.activePeriod == nil {
+		return nil, 0, errors.New("activePeriod must not be nil")
+	}
+	if r.activePeriod != nil && envvar.SourcegraphDotComMode() {
+		return nil, 0, errors.New("site analytics is not available on sourcegraph.com")
+	}
+	r.once.Do(func() {
+		var userIDs []string
+		var err error
+
+		switch *r.activePeriod {
+		case "TODAY":
+			_, userIDs, _, err = useractivity.ListUsersToday()
+		case "THIS_WEEK":
+			_, userIDs, _, err = useractivity.ListUsersThisWeek()
+		case "THIS_MONTH":
+			_, userIDs, _, err = useractivity.ListUsersThisMonth()
+		default:
+			err = fmt.Errorf("unknown user event %s", *r.activePeriod)
+		}
+		if err != nil {
+			r.err = err
+			return
+		}
+
+		r.opt.UserIDs, err = sliceAtoi(userIDs)
+		if err != nil {
+			r.err = err
+			return
+		}
+
+		r.totalCount = len(userIDs)
+		r.users, r.err = db.Users.List(ctx, &r.opt)
+	})
+	return r.users, r.totalCount, r.err
 }
 
 func (r *userConnectionResolver) Nodes(ctx context.Context) ([]*userResolver, error) {
@@ -30,7 +87,13 @@ func (r *userConnectionResolver) Nodes(ctx context.Context) ([]*userResolver, er
 		return nil, err
 	}
 
-	users, err := db.Users.List(ctx, &r.opt)
+	var users []*types.User
+	var err error
+	if r.useCache() {
+		users, _, err = r.compute(ctx)
+	} else {
+		users, err = db.Users.List(ctx, &r.opt)
+	}
 	if err != nil {
 		return nil, err
 	}
@@ -50,8 +113,18 @@ func (r *userConnectionResolver) TotalCount(ctx context.Context) (int32, error) 
 		return 0, err
 	}
 
-	count, err := db.Users.Count(ctx, r.opt)
+	var count int
+	var err error
+	if r.useCache() {
+		_, count, err = r.compute(ctx)
+	} else {
+		count, err = db.Users.Count(ctx, r.opt)
+	}
 	return int32(count), err
+}
+
+func (r *userConnectionResolver) useCache() bool {
+	return r.activePeriod != nil && *r.activePeriod != "ALL_TIME"
 }
 
 // staticUserConnectionResolver implements the GraphQL type UserConnection based on an underlying
@@ -69,3 +142,15 @@ func (r *staticUserConnectionResolver) Nodes() []*userResolver {
 }
 
 func (r *staticUserConnectionResolver) TotalCount() int32 { return int32(len(r.users)) }
+
+func sliceAtoi(sa []string) ([]int32, error) {
+	si := make([]int32, 0, len(sa))
+	for _, a := range sa {
+		i, err := strconv.Atoi(a)
+		if err != nil {
+			return nil, err
+		}
+		si = append(si, int32(i))
+	}
+	return si, nil
+}